Fit LBSPR model to length data

A function that fits the LBSPR model to length data

LBSPRfit(LB_pars = NULL, LB_lengths = NULL, yrs = NA, Control = list(),
  pen = TRUE, verbose = TRUE, useCPP = TRUE, ...)

Arguments

LB_pars

an object of class 'LB_pars' that contains the life history information

LB_lengths

an object of class 'LB_lengths' that contains the length data

yrs

index of years to include. If NA the model is run on all years

Control

a list of control options for the LBSPR model.

pen

apply a penalty if estimate of selectivity is very high?

verbose

display messages?

useCPP

use cpp optimization code?

...

additional parameters to pass to FilterSmooth

Value

a object of class 'LB_obj'

Details

The Control options are:

modtype

Model Type: either Growth-Type-Group Model (default: "GTG") or Age-Structured ("absel")

maxsd

Maximum number of standard deviations for length-at-age distribution (default is 2)

ngtg

Number of groups for the GTG model. Default is 13

P

Proportion of survival of initial cohort for maximum age for Age-Structured model. Default is 0.01

Nage

Number of pseudo-age classes in the Age Structured model. Default is 101

maxFM

Maximum value for F/M. Estimated values higher than this are trunctated to maxFM. Default is 4