Internal function to fit LBSPR model to length data

An internal function that fits the LBSPR model to a single year of length data

LBSPRfit_(yr = 1, LB_pars = NULL, LB_lengths = NULL, Control = list(),
  pen = TRUE, useCPP = TRUE, verbose = TRUE)

Arguments

yr
index of the year column to fit model to

LB_pars
an object of class 'LB_pars' that contains the life history information

LB_lengths
an object of class 'LB_lengths' that contains the length data

Control
a list of control options for the LBSPR model.

pen
apply a penalty if estimate of selectivity is very high?

useCPP
use cpp optimization code?

verbose
display messages?

Value

a object of class 'LB_obj'

Details

The Control options are:

modtype

Model Type: either Growth-Type-Group Model (default: "GTG") or Age-Structured ("absel")

maxsd

Maximum number of standard deviations for length-at-age distribution (default is 2)

ngtg

Number of groups for the GTG model. Default is 13

P

Proportion of survival of initial cohort for maximum age for Age-Structured model. Default is 0.01

Nage

Number of pseudo-age classes in the Age Structured model. Default is 101

maxFM

Maximum value for F/M. Estimated values higher than this are trunctated to maxFM. Default is 4