LBSPR Simulation Model

Function that generates the expected equilbrium size composition given biological parameters, and fishing mortality and selectivity pattern.

LBSPRsim(LB_pars = NULL, Control = list(), verbose = TRUE)

Arguments

LB_pars
an object of class 'LB_pars' that contains the life history information

Control
a list of control options for the LBSPR model.

verbose
display messages?

Value

a object of class 'LB_obj'

Details

The Control options are:

modtype

Model Type: either Growth-Type-Group Model (default: "GTG") or Age-Structured ("absel")

maxsd

Maximum number of standard deviations for length-at-age distribution (default is 2)

ngtg

Number of groups for the GTG model. Default is 13

P

Proportion of survival of initial cohort for maximum age for Age-Structured model. Default is 0.01

Nage

Number of pseudo-age classes in the Age Structured model. Default is 101

maxFM

Maximum value for F/M. Estimated values higher than this are trunctated to maxFM. Default is 4

Examples

 LB_pars <- new("LB_pars")

#> A blank LB_pars object created

#> Default values have been set for some parameters

 LB_pars@MK <- 1.5
 LB_pars@Linf <- 100
 LB_pars@L50 <- 50
 LB_pars@L95 <- 55
 LB_pars@SL50 <- 60
 LB_pars@SL95 <- 65
 LB_pars@FM <- 1
 Sim <- LBSPRsim(LB_pars)

#> BinMax not set. Using default of 1.3 Linf

#> BinMin not set. Using default value of 0

#> BinWidth not set. Using default value of 1/20 Linf

 Sim@SPR

#> [1] 0.5121073