Circle of estimated SPR and target and limit points

A function that creates a circle plot showing the estimated SPR relative to the target and limit reference points

plotSPRCirc(LB_obj = NULL, SPRTarg = 0.4, SPRLim = 0.2,
  useSmooth = TRUE, Title = FALSE, Leg = TRUE, limcol = "#ff1919",
  targcol = "#f2ff02", abtgcol = "#32ff36", labcol = NULL,
  bgcol = "#FAFAFA", labcex = 2, texcex = 1.3)

Arguments

LB_obj
an object of class 'LB_obj' that contains the life history and fishing information

SPRTarg
a numeric value specifying the SPR target

SPRLim
a numeric value specifying the SPR limit

useSmooth
use the smoothed estimates? Usually would want to do this

Title
include the title?

Leg
include the legend?

limcol
colour for SPR Limit (hex; default is red)

targcol
colour for SPR target (hex; default is yellow)

abtgcol
colour for above SPR target (hex; default is green)

labcol
optional fixed colour for estimated SPR label

bgcol
colour for the background

labcex
size for the estimated SPR label

texcex
size for estimated other labels