Plot sampled length structure against target simulated size composition

A function that plots the observed size structure against the expected size composition at the target SPR

plotTarg(LB_pars = NULL, LB_lengths = NULL, yr = 1, Cols = NULL,
  title = NULL, targtext = TRUE, size.axtex = 12, size.title = 14,
  scales = c("fixed", "free_x", "free_y", "free"))

Arguments

LB_pars
an object of class 'LB_pars' that contains the life history and fishing information

LB_lengths
an object of class 'LB_lengths' that contains the observed size data

yr
index for sampled length data (defaults to 1)

Cols
optional character vector of colours for the plot

title
character - optional title for plot

targtext
logical - should the SPR target text be displayed as a subtitle?

size.axtex
size of the axis text

size.title
size of axis title

scales
argument to ggplot2 function. Are scales shared across all facets (the default, "fixed"), or do they vary across rows ("free_x"), columns ("free_y"), or both rows and columns ("free")

Value

a ggplot object